+++ resources/views/components/live/toolbar/groups/style.blade.php
@@ -3,7 +3,7 @@
     @if ($hasColor)
         <button type="button" class="hb-tb__btn hb-tb__color" data-tb-popover="color" aria-haspopup="true" aria-expanded="false" aria-label="Text colour"><b>A</b><i></i></button>
     @endif
-    <button type="button" class="hb-tb__btn" data-tb-action="save" aria-label="Save as block">@include('heisenberg::components.ui.icon', ['name' => 'floppy-disk-back-fill', 'size' => 20])</button>
+    <button type="button" class="hb-tb__btn" data-tb-action="save" aria-label="{{ __('heisenberg::editor.patterns.save_dialog_title') }}">@include('heisenberg::components.ui.icon', ['name' => 'floppy-disk-back-fill', 'size' => 20])</button>
     @if ($hasAlign)
         <button type="button" class="hb-tb__align" data-tb-popover="align" aria-haspopup="true" aria-expanded="false" aria-label="Align">@include('heisenberg::components.ui.icon', ['name' => 'text-align-right-light', 'size' => 20])<span class="hb-tb__cv">@include('heisenberg::components.ui.icon', ['name' => 'caret-down', 'size' => 10])</span></button>
     @endif